one file for both two region and one region
